Prefix filenames of per-run artefacts with an underscore
Since potentially very many per-job artefacts are stored in the log directory, sorting out the artefacts that are per-run could require wading through many files. By giving all per-run artefacts an underscore ("_") at the start of their fiename, alphabetically sorting the files in the log directory becomes sufficient to isolate and review per-run artefacts in the presence of a flood of per-job artefacts.